backend/requirements.txt
asgiref==3.8.1
Django==5.1.4
django-cors-headers==4.6.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
mysqlclient==2.2.6
PyJWT==2.10.1
sqlparse==0.5.3
tzdata==2024.2
